Add support for multiple choice questions with an abitrary number of options.
The difference between the types of questions is specified in the "type" property of the question data object.

The type property can be any of four options: "MCQ", "G_OBJ", "SHORT", "PARAGRAPH".

The previous essay questions are of the "PARAGRAPH" type, the new multiple choice questions are of the "MCQ" type.

Sort questions list before begnning the test to show "MCQ" type questions first, then "G_OBJ", then "SHORT" types and then finally "PARAGRAPH" types at the end. The order of each group of questions should still randomized.

Show the count of each question type with progress bars that show the count of answered so far versus total count for that type.