The last fix didn't work. The content is still not  entirely encapsulated in the card. Here are steps to fix it.
Find the Question display card with classname "flex-grow flex flex-col glass min-h-0"

In it, find the div with classname "p-4 border border-border rounded-md bg-white/50 dark:bg-black/20 flex-grow overflow-auto".

Modify the div to wrap its height to the size of text content set in the paragraph tag with classname "text-lg whitespace-pre-wrap break-words" with a minimum height of 5 lines and maximum of 10 lines, beyond which, it can use the current scolling behaviour.

Modify the card to wrap its height dynamically to encapsulate the question div and the Answer input (MCQ or TextArea).